+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 0.16.0
+## Changed
+- Stopped slideshow in /animes page
+- stopped disqus comments
+- minified js code in /js/app.js
+- optimized logo and images
+- hide the left side of navbar (social media and search button)
+
+## Added
+- google analytics tag
+
+
 # 0.15.0 - 2020-04-26
 ## Changed
 - refactor routes